move ICinderVolumeManager to cinder module.
